fix(shared): restore baseline CI required checks
