+++ .github/workflows/bench.yml
@@ -1,7 +1,8 @@
 name: Benchmark
 
 on:
-  pull_request:
+  schedule:
+    - cron: "0 6 * * 1"
   workflow_dispatch:
 
 jobs:
